requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.3
groq==0.9.0
langchain==0.2.11
langchain_community==0.2.10
langchain_core==0.2.23
langchain_groq==0.1.6
langchain_huggingface==0.0.3
langchain_text_splitters==0.2.2
mage_ai==0.9.72
numpy==2.0.1
pandas==2.2.2
Pillow==10.4.0
PyPDF2==3.0.1
python_docx==1.1.2
Requests==2.32.3
scikit_learn==1.4.2
selenium==4.23.1
streamlit==1.36.0
torch==2.3.1
transformers==4.41.2
~treamlit==1.33.0
~treamlit==1.34.0
